Rename the shared generator module to hegel_gens
The 31 test modules that use it are all called hegel_props.
