+++ geo/src/algorithm/affine_ops.rs
@@ -647,7 +647,7 @@ mod tests {
 #[cfg(test)]
 mod hegel_props {
     use super::{AffineOps, AffineTransform};
-    use crate::utils::pbt_gens::{coords, geometries, star_polygons};
+    use crate::utils::hegel_gens::{coords, geometries, star_polygons};
     use crate::{Coord, MapCoords};
     use hegel::generators::{self, Generator, PrintableGenerator};
